[feat][client] Slice-relative block index with new Slice struct

+++ src/client/vfs/data/slice/common.h
@@ -31,16 +31,16 @@ static std::atomic<uint64_t> slice_seq_id_gen{1};
 
 struct SliceDataContext {
   const uint64_t seq{0};
-  const uint64_t fs_id{0};
+  const uint32_t fs_id{0};
   const uint64_t ino{0};
-  const uint64_t chunk_index{0};
-  const uint64_t chunk_size{0};
-  const uint64_t block_size{0};
-  const uint64_t page_size{0};
+  const int64_t chunk_index{0};
+  const int32_t chunk_size{0};
+  const int32_t block_size{0};
+  const int32_t page_size{0};
 
-  explicit SliceDataContext(uint64_t p_fs_id, uint64_t p_ino,
-                            uint64_t p_chunk_index, uint64_t p_chunk_size,
-                            uint64_t p_block_size, uint64_t p_page_size)
+  explicit SliceDataContext(uint32_t p_fs_id, uint64_t p_ino,
+                            int64_t p_chunk_index, int32_t p_chunk_size,
+                            int32_t p_block_size, int32_t p_page_size)
       : seq(slice_seq_id_gen.fetch_add(1, std::memory_order_relaxed)),
         fs_id(p_fs_id),
         ino(p_ino),
